refactor(BA-7438): move the last-swept-at updater to the v2 bases
`LastSweptAtUpdater` moves to `models/retention/updaters.py` beside the policy
updater already there, on `models/specs/updater.DataUpdater`.

The retention sweep stamps inside its own savepoint, so the legacy
`WriteOps.update_data` now runs the v2 update on that session through
`V2WriteOps`, rather than reaching for a legacy executor.
